feat: Add YAML linting workflow and configuration
This commit adds a new YAML linting workflow and configuration files to the repository. The workflow is triggered on pull requests to the main branch and uses the `ibiqlik/action-yamllint` action to lint YAML files in the `secrets` directory. The linting rules are defined in the `.github/yamllint.yml` configuration file.

Signed-off-by: HAHWUL <hahwul@gmail.com>